preparing for second ISO release

+++ piratepack/src/share/firefox-mods/remove_firefox-mods.sh
@@ -18,10 +18,13 @@ then
     done <"$localdir"/.installed 
 fi
 
-set +e
-chmod -Rf u+rw "$localdir"/* "$localdir"/.[!.]* "$localdir"/...*
-rm -rf "$localdir"/* "$localdir"/.[!.]* "$localdir"/...*
-set -e
+if [ -d "$localdir" ]
+then
+    set +e
+    chmod -Rf u+rw "$localdir"/* "$localdir"/.[!.]* "$localdir"/...*
+    rm -rf "$localdir"/* "$localdir"/.[!.]* "$localdir"/...*
+    set -e
+fi
 
 cd
 
@@ -35,8 +38,23 @@ done < <(find ".mozilla/firefox/"*".default" -maxdepth 0)
 
 cd "$profiledir"
 
-match="$(grep -nr piratelinux.org prefs.js)"
+match="$(grep homepage.*piratelinux.org prefs.js)"
 if [[ "$match" != "" ]]
 then
     echo 'user_pref("browser.startup.homepage", "about:blank");' >> prefs.js
 fi
+
+match="$(grep port.*8124 prefs.js)"
+if [[ "$match" != "" ]]
+then
+    echo 'user_pref("network.proxy.http", "");' >> prefs.js
+    echo 'user_pref("network.proxy.http_port", 0);' >> prefs.js
+    echo 'user_pref("network.proxy.ssl", "");' >> prefs.js
+    echo 'user_pref("network.proxy.ssl_port", 0);' >> prefs.js
+fi
+
+cd
+set +e
+rm -f .local/share/applications/firefox-pm.desktop
+rm -f .local/share/icons/firefox-pm.png
+set -e
